requirements.txt
flask==2.0.1
azure-cosmos==4.2.0
